+++ src/exe.c
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exe.c                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: pboonpro <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/10/18 01:19:09 by pboonpro          #+#    #+#             */
+/*   Updated: 2023/10/18 01:19:09 by pboonpro         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+void	exe(t_exe *a, t_pipe *p_pipe, int *status, char ***env)
+{
+	return ;
+}
